+++ naga/tests/naga/wgsl_errors.rs
@@ -1049,8 +1049,17 @@ macro_rules! check_one_validation {
 ///
 /// NOTE: The only reason we don't use a function for this is because we need to syntactically
 /// re-use `$val_err_pat`.
+///
+/// The optional $other_caps argument at the end specifies capabilities that
+/// allow, the shader or would change the error message if enabled, but do not
+/// get enabled by the specified enable extension. This is only currently the
+/// case for `acceleration_structures` which are enabled by both ray queries
+/// and ray tracing pipelines.
 macro_rules! check_extension_validation {
-    ( $caps:expr, $source:expr, $parse_err:expr, $val_err_pat:pat ) => {
+    ( $caps:expr, $source:expr, $parse_err:expr, $val_err_pat:pat $(, $other_caps:expr)? ) => {
+        #[allow(unused_mut, unused_assignments)]
+        let mut other_caps = naga::valid::Capabilities::empty();
+        $(other_caps = $other_caps;)?
         let caps = $caps;
         let source = $source;
         let mut ext = None;
@@ -1102,7 +1111,9 @@ macro_rules! check_extension_validation {
         };
 
         // Second check, for the expected validation error when the capability is not present
-        let error = naga::valid::Validator::new(naga::valid::ValidationFlags::all(), !caps)
+        // Don't check with explicitly allowed caps, as certain things (currently just
+        // `acceleration_structure`s) can be enabled by multiple extensions
+        let error = naga::valid::Validator::new(naga::valid::ValidationFlags::all(), !(caps | other_caps))
             .validate(&module)
             .map_err(|e| e.into_inner()); // TODO(https://github.com/gfx-rs/wgpu/issues/8153): Add tests for spans
         #[allow(clippy::redundant_pattern_matching)]
@@ -4359,7 +4370,7 @@ fn enumerant_with_template_parameters() {
 }
 
 #[test]
-fn ray_query_enable_extension() {
+fn ray_types_enable_extension() {
     check_extension_validation!(
         Capabilities::RAY_QUERY,
         r#"fn foo() {
@@ -4381,6 +4392,7 @@ fn ray_query_enable_extension() {
         })
     );
 
+    // can be enabled by either of these extensions
     check_extension_validation!(
         Capabilities::RAY_QUERY,
         r#"@group(0) @binding(0)
@@ -4398,7 +4410,28 @@ fn ray_query_enable_extension() {
         Err(naga::valid::ValidationError::Type {
             source: naga::valid::TypeError::MissingCapability(Capabilities::RAY_QUERY),
             ..
-        })
+        }),
+        Capabilities::RAY_TRACING_PIPELINE
+    );
+    check_extension_validation!(
+        Capabilities::RAY_TRACING_PIPELINE,
+        r#"@group(0) @binding(0)
+        var acc_struct: acceleration_structure;
+        "#,
+        r#"error: the `wgpu_ray_query` enable extension is not enabled
+  ┌─ wgsl:2:25
+  │
+2 │         var acc_struct: acceleration_structure;
+  │                         ^^^^^^^^^^^^^^^^^^^^^^ the `wgpu_ray_query` "Enable Extension" is needed for this functionality, but it is not currently enabled.
+  │
+  = note: You can enable this extension by adding `enable wgpu_ray_query;` at the top of the shader, before any other items.
+
+"#,
+        Err(naga::valid::ValidationError::Type {
+            source: naga::valid::TypeError::MissingCapability(Capabilities::RAY_QUERY),
+            ..
+        }),
+        Capabilities::RAY_QUERY
     );
 }
 
@@ -4738,4 +4771,140 @@ fn mesh_shader_enable_extension() {
             ..
         })
     );
+}
+
+/// Checks that every ray tracing pipeline binding in naga is invalid in other stages.
+#[test]
+fn check_ray_tracing_pipeline_bindings() {
+    for (builtin, ty) in [
+        ("ray_invocation_id", "vec3<u32>"),
+        ("num_ray_invocations", "vec3<u32>"),
+        ("instance_custom_data", "u32"),
+        ("geometry_index", "u32"),
+        ("world_ray_origin", "vec3<f32>"),
+        ("world_ray_direction", "vec3<f32>"),
+        ("object_ray_origin", "vec3<f32>"),
+        ("object_ray_direction", "vec3<f32>"),
+        ("ray_t_min", "f32"),
+        ("ray_t_current_max", "f32"),
+        ("object_to_world", "mat4x3<f32>"),
+        ("world_to_object", "mat4x3<f32>"),
+        ("hit_kind", "u32"),
+    ] {
+        for stage in ["@compute @workgroup_size(1)", " @vertex", "@fragment"] {
+            check_one_validation!(
+                &format!(
+                    "{stage}
+            fn main(@builtin({builtin}) v: {ty}) {{}}
+            "
+                ),
+                Err(naga::valid::ValidationError::EntryPoint {
+                    source: naga::valid::EntryPointError::Argument(
+                        0,
+                        naga::valid::VaryingError::InvalidBuiltInStage(_),
+                    ),
+                    ..
+                },)
+            );
+        }
+    }
+}
+
+/// Checks ray generation stage is invalid without enable extension (other stages require `@incoming_payload` which forces a ray payload which is checked in [`check_ray_tracing_pipeline_payload`])
+#[test]
+fn check_ray_tracing_pipeline_ray_generation() {
+    check_extension_validation!(
+            Capabilities::RAY_TRACING_PIPELINE,
+            "@ray_generation
+                fn main() {{}}",
+            "error: the `wgpu_ray_tracing_pipeline` enable extension is not enabled
+  ┌─ wgsl:1:2
+  │
+1 │ @ray_generation
+  │  ^^^^^^^^^^^^^^ the `wgpu_ray_tracing_pipeline` \"Enable Extension\" is needed for this functionality, but it is not currently enabled.
+  │
+  = note: You can enable this extension by adding `enable wgpu_ray_tracing_pipeline;` at the top of the shader, before any other items.
+
+",
+            Err(naga::valid::ValidationError::EntryPoint {
+                source: naga::valid::EntryPointError::UnsupportedCapability(naga::valid::Capabilities::RAY_TRACING_PIPELINE),
+                ..
+            },)
+        );
+}
+
+#[test]
+fn check_ray_tracing_pipeline_payload() {
+    for space in ["ray_payload", "incoming_ray_payload"] {
+        // ascii is a byte per char so length is fine
+        let space_arrows = "^".to_string().repeat(space.len());
+        check_extension_validation!(
+            Capabilities::RAY_TRACING_PIPELINE,
+            &format!("var<{space}> payload: u32;"),
+            &format!("error: the `wgpu_ray_tracing_pipeline` enable extension is not enabled
+  ┌─ wgsl:1:5
+  │
+1 │ var<{space}> payload: u32;
+  │     {space_arrows} the `wgpu_ray_tracing_pipeline` \"Enable Extension\" is needed for this functionality, but it is not currently enabled.
+  │
+  = note: You can enable this extension by adding `enable wgpu_ray_tracing_pipeline;` at the top of the shader, before any other items.
+
+"),
+            Err(naga::valid::ValidationError::GlobalVariable {
+                source: naga::valid::GlobalVariableError::UnsupportedCapability(naga::valid::Capabilities::RAY_TRACING_PIPELINE),
+                ..
+            },)
+        );
+    }
+}
+
+#[test]
+fn check_ray_tracing_pipeline_incoming_payload_required() {
+    for stage in ["any_hit", "closest_hit", "miss"] {
+        // ascii is a byte per char so length is fine
+        let stage_arrows = "^".to_string().repeat(stage.len());
+        check(
+            &format!("enable wgpu_ray_tracing_pipeline; @{stage} fn main() {{}}"),
+            &format!("error: incoming payload is missing on a `closest_hit`, `any_hit` or `miss` shader entry point
+  ┌─ wgsl:1:36
+  │
+1 │ enable wgpu_ray_tracing_pipeline; @{stage} fn main() {{}}
+  │                                    {stage_arrows} must be paired with a `@incoming_payload` attribute
+
+"),
+        );
+    }
+}
+
+#[test]
+fn check_ray_tracing_pipeline_payload_disallowed() {
+    for (stage, output, stmt) in [
+        (
+            "var<incoming_ray_payload> incoming: u32; @any_hit @incoming_payload(incoming)",
+            "",
+            "",
+        ),
+        ("@compute @workgroup_size(1)", "", ""),
+        (
+            "@vertex",
+            " -> @builtin(position) vec4<f32>",
+            "return vec4();",
+        ),
+        ("@fragment", "", ""),
+    ] {
+        check_one_validation!(
+            &format!(
+                "enable wgpu_ray_tracing_pipeline;
+            @group(0) @binding(0) var acc_struct: acceleration_structure;
+            var<ray_payload> payload: u32;
+            
+            {stage} fn main() {output} {{_ = payload; {stmt}}}"
+            ),
+            Err(naga::valid::ValidationError::EntryPoint {
+                source: naga::valid::EntryPointError::RayPayloadInInvalidStage(_),
+                ..
+            },),
+            Capabilities::RAY_TRACING_PIPELINE
+        );
+    }
 }